тестирование дедлайна блет

+++ 2/app/src/main/java/well/keepitsimple/dnevnik/TaskItem.kt
@@ -2,7 +2,7 @@ package well.keepitsimple.dnevnik
 
 data class TaskItem(
     val subject: String,
-    val deadline: Int,
+    val deadline: Long,
     val text: String,
     val id: String,
         )
